conda-requirements.txt
numpy==1.18.1
scipy==1.4.1
pandas==1.0.1
scikit-learn==0.22.1
pytorch==1.4.0
torchvision==0.5.0
ipykernel==5.1.4
scikit-image==0.16.2
wget==1.20.1
